projeto/requirements.txt
flask
python-dotenv
scikit-learn
nltk
textblob
joblib
PyPDF2
transformers
torch
protobuf
sentencepiece
pysentimiento
llama-index
matplotlib
wordcloud
sentence-transformers
deep-translator
# Removido: streamlit, pois não faz mais parte do sistema
